binance_bot/requirements.txt
ccxt
pandas
pyyaml
python-dotenv
requests
schedule
colorlog
optuna
websocket-client
